pyproject.toml
[tool.poetry]
name = "dis_prod_back"
version = "0.1.0"
description = ""
authors = ["juansebastian <jbravomeneses@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
